Step3 리뷰 부탁드립니다. (#480)
* 지하철 구간 등록 인수 테스트 시그니처 메서드 작성

* feat 패키지 구조 변경 및 Section Entity 생성

* feat : 지하철 구간 등록 성공 응답

* feat : 구간 등록 시 구간 상행역 = 노선 하행역

* feat : 에러 메시지 등록

* feat : 공통 응답 별도 분리

* feat : 도메인 컬럼명 변경

* feat : 도메인 구조 변경

* refactor : 공통 부분 step 추출

* error.properties : error message 추가

* feat : 도메인 구조 변경에 따른 response 멤버 변경

* feat : 구간 조회 쿼리 추가

* refactor : 공통 부분 step 추출

* refactor : 도메인 구조 변경에 따른 파라미터 변경

* refactor : 공통 부분 step 추출

* refactor : 구간 삭제 인수 테스트 추가 및 구현

* refactor : 서버_에러_응답 메서드 추출

* refactor : 중복 부분 공통 부분 추출

* refactor : 비개발자가 이해하기 쉬운 인수 테스트 시나리오 작성

* refactor : 도메인 검증 로직 이동 서비스->도메인

* refactor : 미사용 메서드 삭제

* refactor : 구간 일급 콜렉션 사용으로 변경

* refactor : 구간 삭제 로직 도메인 내부로 이동

* feat : lombok 추가

* refactor : 노선 목록 조회시 순환 참조 해결

* feat : objectMapper 사용

* refactor : 불필요 개행 삭제

* refactor : 객체 내부 this 사용

* refactor : error message 변경

* refactor : ObjectUtils 미사용으로 변경

* feat : 구간에 등록된 역 삭제 시 오류 발생

+++ src/main/java/nextstep/subway/applicaion/StationService.java
@@ -1,45 +1,39 @@
 package nextstep.subway.applicaion;
 
+import lombok.RequiredArgsConstructor;
 import nextstep.subway.applicaion.dto.StationRequest;
 import nextstep.subway.applicaion.dto.StationResponse;
-import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
+import nextstep.subway.domain.section.SectionRepository;
+import nextstep.subway.domain.station.Station;
+import nextstep.subway.domain.station.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+@RequiredArgsConstructor
 @Service
 @Transactional(readOnly = true)
 public class StationService {
-    private StationRepository stationRepository;
-
-    public StationService(StationRepository stationRepository) {
-        this.stationRepository = stationRepository;
-    }
+    private final StationRepository stationRepository;
+    private final SectionRepository sectionRepository;
 
     @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
-        Station station = stationRepository.save(new Station(stationRequest.getName()));
-        return createStationResponse(station);
+        return new StationResponse(stationRepository.save(new Station(stationRequest.getName())));
     }
 
     public List<StationResponse> findAllStations() {
-        return stationRepository.findAll().stream()
-                .map(this::createStationResponse)
-                .collect(Collectors.toList());
+        return stationRepository.findAll().stream().map(StationResponse::new).collect(Collectors.toList());
     }
 
     @Transactional
     public void deleteStationById(Long id) {
+        if (sectionRepository.existsByUpStation_Id(id) || sectionRepository.existsByDownStation_Id(id)) {
+            throw new IllegalArgumentException("station.used.section");
+        }
         stationRepository.deleteById(id);
     }
 
-    private StationResponse createStationResponse(Station station) {
-        return new StationResponse(
-                station.getId(),
-                station.getName()
-        );
-    }
 }
